+++ react-app/src/components/SearchBar/SearchBar.css
@@ -32,3 +32,14 @@
   height: 40px;
   width: auto;
 }
+
+#datebox {
+  font-family: Airbnb;
+  border-radius: 32px;
+  border: none;
+  background-color: rgba(255, 255, 255, 0.0);
+}
+
+.nice-dates-day {
+  font-family: Airbnb;
+}
